Move ConvergenceCriteria tag to Convergence::Tags
This tag can be used by other iterative algorithms, in particular the
upcoming nonlinear solver.
