+++ tests/Unit/NumericalAlgorithms/Convergence/Test_Tags.cpp
@@ -17,6 +17,11 @@ struct TestLabel {};
 
 SPECTRE_TEST_CASE("Unit.Numerical.Convergence.Tags",
                   "[Unit][NumericalAlgorithms]") {
+  TestHelpers::db::test_simple_tag<
+      Convergence::Tags::Criteria<TestLabel>>(
+      "ConvergenceCriteria(TestLabel)");
+  TestHelpers::db::test_simple_tag<Convergence::Tags::Iterations<TestLabel>>(
+      "Iterations(TestLabel)");
   TestHelpers::db::test_simple_tag<Convergence::Tags::IterationId<TestLabel>>(
       "IterationId(TestLabel)");
   TestHelpers::db::test_simple_tag<Convergence::Tags::HasConverged<TestLabel>>(
