+++ tests/Unit/ParallelAlgorithms/LinearSolver/Test_Tags.cpp
@@ -45,12 +45,6 @@ SPECTRE_TEST_CASE("Unit.ParallelAlgorithms.LinearSolver.Tags",
   TestHelpers::db::test_prefix_tag<LinearSolver::Tags::Preconditioned<Tag>>(
       "Preconditioned(Tag)");
   TestHelpers::db::test_simple_tag<
-      LinearSolver::Tags::ConvergenceCriteria<TestOptionsGroup>>(
-      "ConvergenceCriteria(TestLinearSolver)");
-  TestHelpers::db::test_simple_tag<
-      LinearSolver::Tags::Iterations<TestOptionsGroup>>(
-      "Iterations(TestLinearSolver)");
-  TestHelpers::db::test_simple_tag<
       LinearSolver::Tags::Verbosity<TestOptionsGroup>>(
       "Verbosity(TestLinearSolver)");
